// src/interfaces/race.ts
export interface Race {
  race: string;
  mainProfile: number[];
  secondaryProfile: number[];
  startingWounds: number[];
  startingFatePoints: number[];
  skills: string[];
  talents: string[];
  height: number;
  weight: number[];
  hairColour: string[];
  eyeColour: string[];
  numberOfSiblings: number[];
  age: number[];
  birthPlace: string[];
  femaleName: string[];
  maleName: string[];
}